Updated wro4j library to fix build failure.
wro4j plugin (that compresses JS) was using a buggy library that
was failing when there were spaces in the path to one of its dependencies.
After in Jenkins Build job a private maven repo was switched on, its
the repo is located in a folder with spaces:
`JC. Unit Tests/workspace/.repository/org/webjars/jslint/1d8c1f8f74/...`.
Because of that Jenkins Build Job started to fail. The bug was fixed in
1.7.1 verion of wro4j while our was 1.7.0. Updating to 1.7.5 because
it's the latest.

Details of the wro4j bug:
https://code.google.com/p/wro4j/issues/detail?id=746