JC-2073 Add VoteUp/VoteDown tooltips
It's important that Vote-related tooltips & labels encourage
high-quality questions & answers. Otherwise people will press Vote Up
just because they like the Q or A, not because they are useful.
Or vice versa.

Also, Bootstrap's tooltip usage was updated - it was listening for
`a`, `span` and other elements. But the off doc recommends using
`data-toggle="tooltip"` attributes for elements to have styled
tooltips. Old selectors didn't change because it would be a very big
task to fix all the tooltips in the whole forum, for now let them be.