JC-1269 Available delete button for user without rights fix
For making delete button available on the first post there was a
check that logined user has rights to delete his own posts. Also,there
was check that there are only posts of the author of the first post.
But there was not verification that currently logined user is actually
the author of this post. I've added this check.