JC-2071 Additional incorrect tooltip over the button "Upload"
- fixed the style and behavior of "input[type='file']".
- added an event handler "onCkick" to the element "a#upload".
- overriden method of the upload button.

Comments to the commit:
For the text of tooltips in the item "input[type=file]" meets the
"title" attribute. If it is empty, the tooltip does not appear. But there
is a problem is that in different browsers under empty means different
(in some it is "", and in others it is " ").
Thus, once the tooltip can be removed in two ways:
1. replace the "title" attribute of element "input.
2. move element "input[type=file]" beyond the wrapper element (in our case
   it "a id='upload'") to the hover event of the mouse cursor did not
   begin with the element "input". Further, the event "click" of the
   element "input" to trigger using the event "click" of the wrapper element
   and its contained elements.
This commit is an attempt to implement the second option.