JC-2278 Fix caching tests that depended on locale
The dates, when converted into string, depend a lot on the locale. If
the locale is Russian, then the date will be translated into russian.
The date that browsers use is in US format, that's why we use this
locale when we transform date from HTTP headers with `Locale.US`. So
tests need to use the same local to make this possible.