JC-2170 Wrong tags for sintax highlight in Opera12
Some browsers could have selectionStart and selectionEnd that is not equal
to 0 for empty textarea.
E.g. opera 12 has 15. So it's better to assign 0 in order not to have
problems in the future.
I also incidentally found and corrected an inaccuracy in the file
bbeditor.tag.
It was
  title='<spring:message code='label.answer.highlight'/>'>
changed to
  title="<spring:message code='label.answer.highlight'/>">