JC-2309 Notifications bug fix
I fixed the problem with checkbox email notifications.
The cause of this problem is was a comparing admin value 'emailNotification' as a boolean(=== true),
but need to compare this as a String(=== 'true') because the return value
type of 'emailNotification' is a String.
Actual result if email notifications is disabled - checkbox is 'unchecked'