Include performance tests module into reactor
We didn't include Performance System Tests in the reactor because we
didn't want to run them when unit/component tests are run. But the
downside was that you can't run the performance tests without
previously installing the parent POM.

To have both problems solved - the module is added to the list of
modules, but `mvn test` won't invoke the actual tests until a special
flag is passed. Ultimately to run just the performance tests:
`mvn -pl jcommune-performance-tests -Dperformance-test.skip=false`