JC-2350 Fixed: smileys (UTF-8) are displayed as rombuses in title of the question
JC-2351 Fixed: smileys (UTF-8) are displayed as rombuses
in comments to the question

Fixed not correct escaping of Emoji in other elements of user
interface of QuestionsAndAnswers plugin.

To solve tasks used Unbescape Java library for correct escaping
strings that contain Emoji:  https://github.com/unbescape/unbescape

Next step: writing tests to check the correctness of text and title
in questions and answers.